refactor: convert to tsx + add custom props
